Components Folder Restructure
