+++ front-end/app/components/HomePage/HeroCarouselHomePage/carouselComponent.module.css
