+++ front-end/app/components/HomePage/HeroListCarousel/index.tsx
@@ -2,8 +2,9 @@ import Link from 'next/link'
 import React from 'react'
 import Image from 'next/image'
 import styles from "./component.module.css"
+import { ApiDefaultResult } from '@/app/ts/interfaces/apiAnilistDataInterface'
 
-function ListCarousel({ data }: any) {
+function ListCarousel({ data }: {data: ApiDefaultResult}) {
     return (
         <li className={styles.container}>
             <Link href={`/media/${data.id}`}>
