+++ front-end/app/layout/header/components/SearchResultItemCard/index.tsx
